moved media-selection-stor into stores folder of media-bundle
