+++ src/Sulu/Bundle/MediaBundle/Resources/js/stores/MediaSelectionStore/MediaSelectionStore.js
@@ -1,9 +1,9 @@
 // @flow
 import {action, computed, observable} from 'mobx';
 import type {IObservableValue} from 'mobx'; // eslint-disable-line
-import {arrayMove} from 'sulu-admin-bundle/components';
-import {ResourceRequester} from 'sulu-admin-bundle/services';
-import type {MediaItem} from '../types';
+import {arrayMove} from 'sulu-admin-bundle/components/index';
+import {ResourceRequester} from 'sulu-admin-bundle/services/index';
+import type {MediaItem} from '../../containers/MediaSelection/types';
 
 const THUMBNAIL_SIZE = 'sulu-25x25';
 const MEDIA_RESOURCE_KEY = 'media';
